Added limits for data client can enter in when signing to the website.

+++ src/main/java/ee/taltech/iti03022023salonbackend/service/ClientService.java
@@ -86,25 +86,81 @@ public class ClientService {
      * Method for adding a new client to the database.
      * Because of every client has to have their unique email address, function will check
      * if the database doesn't already have a client with the same email.
+     * <p>
+     * Returns the status number which explain the result:
+     * 0 - correct and a new user is registered
+     * 1 - client already exists in the database
+     * 2 - password already in use
+     * 3 - unacceptable password
+     * 4 - unacceptable phone number
+     * 5 - unacceptable id code
      *
      * @param client to be added
-     * @return the string explaining the result
+     * @return status number
      */
     @Transactional
     public String addClient(Client client, String password) {
         Optional<Client> existingClient = clientRepository.findClientsByEmailIgnoreCase(client.getEmail());
         Optional<User> existingUser = userRepository.findUsersByPasswordIgnoreCase(password);
         if (existingClient.isPresent()) {
-            return "Client already exists in the database.";
+            return "1";
         } else if (existingUser.isPresent()) {
-            return "Password already in use.";
+            return "2";
+        } else if (isValidPassword(password)) {
+            return "3";
+        } else if (isValidPhoneNumber(client.getPhoneNumber())) {
+            return "4";
+        } else if (isValidIdCode(client.getIdCode())) {
+            return "5";
         }
+        client.setMoney(0);
         clientRepository.save(client);
         User newUser = new User();
         newUser.setPassword(password);
         newUser.setClientId(client.getClientId());
         userRepository.save(newUser);
-        return client.getFirstName() + " " + client.getLastName();
+        return "0";
+    }
+
+    /**
+     * Control method to check password's validity.
+     *
+     * @param password to be controlled
+     * @return boolean
+     */
+    public boolean isValidPassword(String password) {
+        String alphabet = "abcdefghijklmnopqrstuvwxyz";
+        Integer counter = 0;
+        if (password.length() < 8 || password.length() > 50) {
+            return false;
+        } else if (password.equals(password.toLowerCase())) {
+            return false;
+        }
+        for (int i = 0; i < password.length(); i++) {
+            if (alphabet.contains(password.substring(i, i + 1))) {
+                counter++;
+            }
+        }
+        return !counter.equals(password.length());
+    }
+
+    /**
+     * Control method to check validity of phone number.
+     *
+     * @param number to be controlled
+     * @return boolean
+     */
+    public boolean isValidPhoneNumber(String number) {
+        String numerics = "1234567890";
+        if (number.length() < 7) {
+            return false;
+        }
+        for (int i = 0; i < numerics.length(); i++) {
+            if (!numerics.contains(number.substring(i, i + 1))) {
+                return false;
+            }
+        }
+        return true;
     }
 
     /**
@@ -127,6 +183,66 @@ public class ClientService {
         }
         User user = existingUser.get();
         return user.getPassword().equals(password);
+    }
+
+    /**
+     * Method for checking ID code's validity.
+     *
+     * @param idCode to be checked
+     * @return boolean
+     */
+    public boolean isValidIdCode(String idCode) {
+        List<Integer> months1 = List.of(1, 3, 5, 7, 8, 10, 12);
+        List<Integer> months2 = List.of(4, 6, 9, 11);
+        int genderNumber = Integer.parseInt(idCode.substring(0, 1));
+        int yearNumber = Integer.parseInt(idCode.substring(1, 3));
+        int monthNumber = Integer.parseInt(idCode.substring(3, 5));
+        int dayNumber = Integer.parseInt(idCode.substring(5, 7));
+        if (idCode.length() != 11) {
+            return false;
+        } else if (genderNumber > 6 || genderNumber < 1) {
+            return false;
+        } else if (yearNumber < 0 || yearNumber > 99) {
+            return false;
+        } else if (monthNumber < 1 || monthNumber > 12) {
+            return false;
+        } else if (dayNumber > 31 || dayNumber < 1) {
+            return false;
+        } else if (dayNumber <= 31 && !months1.contains(monthNumber)) {
+            return false;
+        } else if (dayNumber <= 30 && !months2.contains(monthNumber)) {
+            return false;
+        } else return isControlNumberCorrect(idCode);
+    }
+
+    /**
+     * Method for checking control number validity.
+     *
+     * @param idCode to be checked
+     * @return boolean
+     */
+    public boolean isControlNumberCorrect(String idCode) {
+        int one = Integer.parseInt(idCode.substring(0, 1));
+        int two = Integer.parseInt(idCode.substring(1, 2));
+        int three = Integer.parseInt(idCode.substring(2, 3));
+        int four = Integer.parseInt(idCode.substring(3, 4));
+        int five = Integer.parseInt(idCode.substring(4, 5));
+        int six = Integer.parseInt(idCode.substring(5, 6));
+        int seven = Integer.parseInt(idCode.substring(6, 7));
+        int eight = Integer.parseInt(idCode.substring(7, 8));
+        int nine = Integer.parseInt(idCode.substring(8, 9));
+        int ten = Integer.parseInt(idCode.substring(9, 10));
+        int last = Integer.parseInt(idCode.substring(10, 11));
+        int first_algo = one + two * 2 + three * 3 + four * 4 + five * 5 + six * 6 + seven * 7 + eight * 8 + nine * 9 + ten;
+        int second_algo = one * 3 + two * 4 + three * 5 + four * 6 + five * 7 + six * 8 + seven * 9 + eight + nine * 2 + ten * 3;
+        if (first_algo % 11 == last) {
+            return true;
+        } else if (second_algo % 11 == last && first_algo % 11 >= 10) {
+            return true;
+        } else if (second_algo % 11 >= 10 && last == 0) {
+            return true;
+        }
+        return false;
     }
 
     /**
@@ -158,6 +274,12 @@ public class ClientService {
             return "No such client in the database.";
         }
         Client client = existingClient.get();
+        Optional<User> existingUser = userRepository.findUsersByClientId(client.getClientId());
+        if (existingUser.isEmpty()) {
+            return "Error";
+        }
+        User user = existingUser.get();
+        userRepository.delete(user);
         clientRepository.delete(client);
         return "Client " + client.getFirstName() + " " + client.getLastName() + " was removed from the database.";
     }
@@ -182,6 +304,21 @@ public class ClientService {
             }
         }
         return salonServiceDtoList;
+    }
+
+    /**
+     * Method for adding money to client's account.
+     *
+     * @param clientId to find client from database
+     * @param amount to be added
+     */
+    public void addMoney(Long clientId, Integer amount) {
+        Optional<Client> existingClient = clientRepository.findById(clientId);
+        if (existingClient.isPresent()) {
+            Client client = existingClient.get();
+            client.setMoney(client.getMoney() + amount);
+            clientRepository.save(client);
+        }
     }
 
     /**
